Changed the local Charge test to match the new flow
